Implement routing for all tabs
