+++ Views/Communication.php
@@ -1,9 +1,9 @@
 <div class="text-center">
-  <form method="post" class="" locale="ENGLISH" action="./?page=1">
-    <div id="parentForm">
+  <form method="post" class="" locale="ENGLISH" action="./?page=5">
+    <div id="communicationForm">
 
-      <div id="page0" class="container card p-4">
-        <h1>Parent A</h1>
+      <div id="page4" class="container card p-4">
+        <h1>COMMUNICATION FORM</h1>
 
         <div class="form-group row p-3">
           <div class="col-lg-6">
