+++ Views/Header.php
@@ -38,12 +38,12 @@
             </div>
             <ul class="navbar-nav text-center">
                 <li class="nav-item"><a class="nav-link" href="./?page=0">Section 1: <br /> Parties </a></li>
-                <li class="nav-item"><a class="nav-link" href="/coparentingplanpro/?page=1"> Section 2: <br /> Children </a></li>
-                <li class="nav-item"><a class="nav-link"> Section 3: <br /> Legal Custody </a></li>
-                <li class="nav-item"><a class="nav-link" href="./?page=2"> Section 4: <br /> Physical Custody and Timesharing </a></li>
-                <li class="nav-item"><a class="nav-link"> Section 5: <br /> Communication </a></li>
-                <li class="nav-item"><a class="nav-link"> Section 6: <br /> Support of the Child(ren) </a></li>
-                <li class="nav-item"><a class="nav-link"> Section 7: <br /> Other Issues </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=1"> Section 2: <br /> Children </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=2"> Section 3: <br /> Legal Custody </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=3"> Section 4: <br /> Physical Custody and Timesharing </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=4"> Section 5: <br /> Communication </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=5"> Section 6: <br /> Support of the Child(ren) </a></li>
+                <li class="nav-item"><a class="nav-link" href="./?page=6"> Section 7: <br /> Other Issues </a></li>
                 <li class="nav-item"><a class="nav-link"> Section 8: <br /> Legal </a></li>
             </ul>
         </nav>
